+++ test/unit_problem_solving_searching.lua
@@ -34,6 +34,7 @@ require("depth_limited_search")
 require("iterative_deepening_search")
 require("greedy_search")
 require("a_star_search")
+require("recursive_best_first_search")
 
 print("Breadth-first search")
 print("--------------------")
@@ -47,10 +48,15 @@ print("Solution is Munchen...",
 
 print("Path cost is... " .. solution.cost)
 
-print("Parent 1 is Frankfurt...", 
-  solution.parent[1] == germany.state.Frankfurt)
-print("Parent 2 is Kassel...", 
-  solution.parent[2] == germany.state.Kassel)
+for i = 1, #solution.parent do
+  local cName = nil
+  for city,id in pairs(germany.state) do
+    if (id == solution.parent[i]) then
+      print("Parent " .. i .. ": " .. city)
+      break
+    end
+  end
+end
 
 print("Elapsed: " .. t2 - t1)
 
@@ -68,12 +74,15 @@ print("Solution is Munchen...",
 
 print("Path cost is... " .. solution.cost[1])
 
-print("Parent 1 is Frankfurt...", 
-  solution.parent[1] == germany.state.Frankfurt)
-print("Parent 2 is Wurzburg...", 
-  solution.parent[2] == germany.state.Wurzburg)
-print("Parent 3 is Nurnberg...", 
-  solution.parent[3] == germany.state.Nurnberg)
+for i = 1, #solution.parent do
+  local cName = nil
+  for city,id in pairs(germany.state) do
+    if (id == solution.parent[i]) then
+      print("Parent " .. i .. ": " .. city)
+      break
+    end
+  end
+end
 
 print("Elapsed: " .. t2 - t1)
 
@@ -91,8 +100,15 @@ print("Solution is Munchen...",
 
 print("Path cost is... " .. solution.cost)
 
-print("Parent 1 is Frankfurt...", 
-  solution.parent[1] == germany.state.Frankfurt)
+for i = 1, #solution.parent do
+  local cName = nil
+  for city,id in pairs(germany.state) do
+    if (id == solution.parent[i]) then
+      print("Parent " .. i .. ": " .. city)
+      break
+    end
+  end
+end
 
 print("Elapsed: " .. t2 - t1)
 
@@ -184,6 +200,33 @@ print("-------------")
 t1 = os.clock()
 solution = a_star_search(germany.problem, germany.state.Frankfurt,
   germany.state.Munchen, true, 
+  {378, 341, 265, 216, 282, 379, 150, 63, 0, 458})
+t2 = os.clock()
+
+print("Solution is Munchen...",
+  solution.state == germany.state.Munchen)
+
+print("Path cost is... " .. solution.cost[1])
+
+for i = 1, #solution.parent do
+  local cName = nil
+  for city,id in pairs(germany.state) do
+    if (id == solution.parent[i]) then
+      print("Parent " .. i .. ": " .. city)
+      break
+    end
+  end
+end
+
+print("Elapsed: " .. t2 - t1)
+
+print("")
+
+print("Recursive best-first search")
+print("---------------------------")
+t1 = os.clock()
+solution = recursive_best_first_search(germany.problem,
+  germany.state.Frankfurt, germany.state.Munchen,
   {378, 341, 265, 216, 282, 379, 150, 63, 0, 458})
 t2 = os.clock()